+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changes that happened in releases
 * Fix issue with unions in GraphQL selection set
 * Fix issue with `EMBEDDING_VIEW` macro when used in subquery within batch loaded mapping
 * Fix issue with resolving function return types in Hibernate 6 integration
+* Fix fetch determination for relay node types
 
 ### Backwards-incompatible changes